优化折线图；调试控件位置

+++ Hpi-FDS/NTShipCompanyTranShareDao.m
@@ -173,6 +173,13 @@ static sqlite3 *database;
 
 
 +(void) InsertByPortCode:(NSMutableArray *)portCode :(NSString *)startDate :(NSString *)endDate{
+    
+    char *errorMsg;
+    if (sqlite3_exec(database, "BEGIN;", 0, 0, &errorMsg)!=SQLITE_OK) {
+        sqlite3_close(database);
+        NSLog(@"exec begin error");
+        return;
+    }
     [NTShipCompanyTranShareDao deleteAll_tmpTable];
     NSMutableString *tmpString = [[NSMutableString alloc] init ];
     NSInteger   sumLW=0;
@@ -203,7 +210,7 @@ static sqlite3 *database;
     for (int i=0; i<monthNum; i++) {
         sqlite3_stmt *statement;
         NSString *sql=[NSString stringWithFormat:@"SELECT sum(lw) from NTShipCompanyTranShare where tradeyear='%@' and trademonth='%@' %@ ",year,month,tmpString];
-       NSLog(@"执行 InsertByPortCode Sql[%@] ",sql);
+//       NSLog(@"执行 InsertByPortCode Sql[%@] ",sql);
         
         if(sqlite3_prepare_v2(database,[sql UTF8String],-1,&statement,NULL)==SQLITE_OK){
             while (sqlite3_step(statement)==SQLITE_ROW) {
@@ -215,8 +222,8 @@ static sqlite3 *database;
         if (sumLW>0) {
             
             sql=[NSString stringWithFormat:@"select comid,company,tradeyear,trademonth,sum(lw) from NTShipCompanyTranShare where tradeyear='%@' and trademonth='%@' %@ group by comid,company,tradeyear,trademonth",year,month,tmpString];
-            NSLog(@"执行 InsertByPortCode Sql[%@] ",sql);
-            if(sqlite3_prepare_v2(database,[sql UTF8String],-1,&statement,NULL)==SQLITE_OK){
+//            NSLog(@"执行 InsertByPortCode Sql[%@] ",sql);
+            if(sqlite3_prepare_v2(database,[sql     UTF8String],-1,&statement,NULL)==SQLITE_OK){
                 while (sqlite3_step(statement)==SQLITE_ROW) {
                     
                     
@@ -247,7 +254,7 @@ static sqlite3 *database;
                     
                     float percent=(float)ntShipCompanyTranShare.LW/sumLW;
                     //保留三位小数
-                    NSLog(@"%0.3f",percent);
+//                    NSLog(@"%0.3f",percent);
                     ntShipCompanyTranShare.PERCENT =[NSString stringWithFormat:@"%0.1f", percent*100];
 
                     [NTShipCompanyTranShareDao insert_tmpTable:ntShipCompanyTranShare];
@@ -272,7 +279,12 @@ static sqlite3 *database;
         }
         sqlite3_finalize(statement);
 
-    }    
+    }
+    if (sqlite3_exec(database, "COMMIT;", 0, 0, &errorMsg)!=SQLITE_OK) {
+        sqlite3_close(database);
+        NSLog(@"exec commit error");
+        return;
+    }
     NSLog(@"insert over");
     [tmpString release];
 }
@@ -349,8 +361,7 @@ static sqlite3 *database;
 	else
 	{
 //		NSLog(@"update success");
-		
-	}
+    }
 	return;
 }
 @end
